[nfc] moving jfc definitions around
I had to do it in my PR to make things compile. Extracting to a separate
change to reduce the noise level.
