+++ src/workerd/jsg/meta.h
@@ -1,0 +1,58 @@
+// Copyright (c) 2017-2022 Cloudflare, Inc.
+// Licensed under the Apache 2.0 license found in the LICENSE file or at:
+//     https://opensource.org/licenses/Apache-2.0
+
+#pragma once
+// INTERNAL IMPLEMENTATION FILE
+
+#include <kj/tuple.h>
+
+namespace workerd::jsg {
+
+class Lock;
+
+template <typename T>
+struct ArgumentIndexes_;
+template <typename T, typename Ret, typename... Args>
+struct ArgumentIndexes_<Ret (T::*)(Args...)> {
+  using Indexes = kj::_::MakeIndexes<sizeof...(Args)>;
+};
+template <typename T, typename Ret, typename... Args>
+struct ArgumentIndexes_<Ret (T::*)(Lock&, Args...)> {
+  using Indexes = kj::_::MakeIndexes<sizeof...(Args)>;
+};
+template <typename T, typename Ret, typename... Args>
+struct ArgumentIndexes_<Ret (T::*)(const v8::FunctionCallbackInfo<v8::Value>&, Args...)> {
+  using Indexes = kj::_::MakeIndexes<sizeof...(Args)>;
+};
+template <typename T, typename Ret, typename... Args>
+struct ArgumentIndexes_<Ret (T::*)(Args...) const> {
+  using Indexes = kj::_::MakeIndexes<sizeof...(Args)>;
+};
+template <typename T, typename Ret, typename... Args>
+struct ArgumentIndexes_<Ret (T::*)(Lock&, Args...) const> {
+  using Indexes = kj::_::MakeIndexes<sizeof...(Args)>;
+};
+template <typename T, typename Ret, typename... Args>
+struct ArgumentIndexes_<Ret (T::*)(const v8::FunctionCallbackInfo<v8::Value>&, Args...) const> {
+  using Indexes = kj::_::MakeIndexes<sizeof...(Args)>;
+};
+template <typename Ret, typename... Args>
+struct ArgumentIndexes_<Ret(Args...)> {
+  using Indexes = kj::_::MakeIndexes<sizeof...(Args)>;
+};
+template <typename Ret, typename... Args>
+struct ArgumentIndexes_<Ret(Lock&, Args...)> {
+  using Indexes = kj::_::MakeIndexes<sizeof...(Args)>;
+};
+template <typename Ret, typename... Args>
+struct ArgumentIndexes_<Ret(const v8::FunctionCallbackInfo<v8::Value>&, Args...)> {
+  using Indexes = kj::_::MakeIndexes<sizeof...(Args)>;
+};
+template <typename T>
+using ArgumentIndexes = typename ArgumentIndexes_<T>::Indexes;
+// ArgumentIndexes<SomeMethodType> expands to kj::_::Indexes<0, 1, 2, 3, ..., n-1>, where n is the
+// number of arguments to the method, not counting the magic Lock or FunctionCallbackInfo parameter
+// (if any).
+
+}  // namespace workerd::jsg
